requirements.txt
python-dotenv
httpx
pillow
pymysql
